Move getLabel/getSublabel to wpcom-checkout package
